rework solution from a marker based one to a directory separated
