+++ .github/workflows/version-override-test.yaml
@@ -5,7 +5,6 @@ on:
 jobs:
   integration:
     uses: agntcy/coffeeAgntcy/.github/workflows/test.yaml@integration-hook
-    secrets: inherit
     with:
       pip_overrides: |
         httpx==0.27.2
