+++ coffeeAGNTCY/coffee_agents/corto/tests/integration_llm/conftest.py
@@ -1,0 +1,4 @@
+# Copyright AGNTCY Contributors (https://github.com/agntcy)
+# SPDX-License-Identifier: Apache-2.0
+
+pytest_plugins = ["tests.integration.conftest"]
