+++ coffeeAGNTCY/coffee_agents/lungo/pyproject.toml
@@ -89,9 +89,6 @@ filterwarnings = [
 ]
 markers = [
     "agents(names): marks tests that run for specific farms (e.g. @pytest.mark.farms([\"brazil\"]))",
-    "live_server: subprocess uvicorn/A2A HTTP; in-process TestClient insufficient",
-    "docker: requires docker-compose session fixtures",
-    "llm: requires live LiteLLM/Azure credentials",
     "no_pin_auction_shared: do not keep agents.supervisors.auction.graph.shared across this test",
 ]
 pythonpath = ["."]
