+++ coffeeAGNTCY/coffee_agents/lungo/tests/integration_llm/test_pattern_chat_proxy.py
@@ -1,7 +1,7 @@
 # Copyright AGNTCY Contributors (https://github.com/agntcy)
 # SPDX-License-Identifier: Apache-2.0
 
-"""Live smoke test for POST /patterns/{name}/chat against a real LLM proxy.
+"""Smoke test for POST /patterns/{name}/chat against a real LiteLLM proxy.
 
 Spins a short-lived uvicorn subprocess (TestClient buffers streaming responses
 and breaks ADK's async LiteLLM client), then issues a real request against the
@@ -19,26 +19,23 @@ from urllib.parse import quote
 
 import httpx
 import pytest
-from tests.helpers.workflow_api_auth import workflow_api_auth_headers
-from tests.unit.agentic_workflows.api.agentic_uvicorn_helpers import (
+
+from tests.helpers.agentic_uvicorn_helpers import (
     assert_lungo_package_layout,
     free_tcp_port,
     start_agentic_uvicorn,
     wait_health,
 )
+from tests.helpers.workflow_api_auth import workflow_api_auth_headers
 
 
-pytestmark = [
-    pytest.mark.skipif(
-        not (os.getenv("LITELLM_PROXY_BASE_URL") and os.getenv("LITELLM_PROXY_API_KEY")),
-        reason="LITELLM_PROXY_BASE_URL / LITELLM_PROXY_API_KEY not configured",
-    ),
-    pytest.mark.live_server,
-    pytest.mark.llm,
-]
+pytestmark = pytest.mark.skipif(
+    not (os.getenv("LITELLM_PROXY_BASE_URL") and os.getenv("LITELLM_PROXY_API_KEY")),
+    reason="LITELLM_PROXY_BASE_URL / LITELLM_PROXY_API_KEY not configured",
+)
 
 
-def test_live_pattern_chat_streams_real_response() -> None:
+def test_pattern_chat_proxy_streams_real_response() -> None:
     assert_lungo_package_layout()
     port = free_tcp_port()
     base_url = f"http://127.0.0.1:{port}"
