+++ coffeeAGNTCY/coffee_agents/lungo/tests/live/test_workflow_instance_sse_http_stream.py
@@ -27,7 +27,7 @@ import pytest
 from schema.types import Event
 
 from tests.helpers.workflow_api_auth import workflow_api_auth_headers
-from tests.unit.agentic_workflows.api.agentic_uvicorn_helpers import (
+from tests.helpers.agentic_uvicorn_helpers import (
     assert_lungo_package_layout,
     first_sse_data_payload,
     free_tcp_port,
@@ -38,7 +38,6 @@ from tests.unit.agentic_workflows.api.agentic_uvicorn_helpers import (
 )
 
 
-@pytest.mark.live_server
 def test_sse_stream_receives_event_after_post() -> None:
     assert_lungo_package_layout()
     port = free_tcp_port()
