+++ coffeeAGNTCY/coffee_agents/recruiter/tests/integration_llm/test_caching.py
@@ -3,7 +3,7 @@
 
 """Integration tests for response and tool caching.
 
-Run with: pytest tests/integration/test_caching.py -v
+Run with: pytest tests/integration_llm/test_caching.py -v
 """
 
 import asyncio
@@ -11,9 +11,6 @@ import time
 
 import pytest
 from agent_recruiter.recruiter import RecruiterTeam
-
-
-pytestmark = pytest.mark.llm
 
 
 @pytest.fixture
